server/requirements.txt
Flask==2.3.3
flask-restful==0.3.10
flask-cors==4.0.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
PyYAML==6.0.1
